swiftlint compliance

+++ CavEat/Services/DataManager.swift
@@ -11,19 +11,20 @@ import UIKit
 import CoreData
 
 class DataManager {
-  static let shared = DataManager(context: nil)
-  
+    static let shared = DataManager(context: nil)
+
     let context: NSPersistentContainer
-    
+
     init(context: NSPersistentContainer?) {
       if let context = context {
         self.context = context
       } else {
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
+        let appDelegate = UIApplication.shared.delegate as! AppDelegate // swiftlint:disable:this force_cast
         self.context = appDelegate.persistentContainer
       }
     }
-  
+
+    // swiftlint:disable function_body_length
     func saveFood(food: Food) -> Bool {
         if food.createdAt != nil { // update the food
             // fetch the specific food
@@ -33,7 +34,7 @@ class DataManager {
             request.predicate = NSPredicate(format: "id = %@", id as CVarArg)
             do {
                 let result = try context.viewContext.fetch(request)
-                let foodDataList  = result as! [NSManagedObject]
+                guard let foodDataList = result as? [NSManagedObject] else { return false }
                 guard foodDataList.count > 0 else { return false }
                 let foodData = foodDataList[0]
                 // realistically we will only ever update the name
@@ -49,36 +50,36 @@ class DataManager {
                 } else {
                     newFood.setValue(UUID(), forKey: "id")
                 }
-                newFood.setValue(food.api_id, forKey: "api_id")
+                newFood.setValue(food.apiId, forKey: "api_id")
                 newFood.setValue(food.name, forKey: "name")
                 newFood.setValue(food.upc, forKey: "upc")
                 newFood.setValue(Date(), forKey: "created_at")
                 // Loop over ingredients and nutritionFacts and add to array
                 // Should refactor these into helpers
                 let ingredients = newFood.mutableSetValue(forKey: #keyPath(CD_food.ingredients))
-                for i in food.ingredients {
+                for ingredient in food.ingredients {
                     if let ingEntity = NSEntityDescription.entity(forEntityName: "CD_ingredient", in: context.viewContext) {
                         let newIng = NSManagedObject(entity: ingEntity, insertInto: context.viewContext)
-                        newIng.setValue(i.id, forKey: "id")
-                        newIng.setValue(i.name, forKey: "name")
-                        newIng.setValue(i.composition, forKey: "composition")
-                        newIng.setValue(i.description, forKey: "ing_description")
-                        newIng.setValue(i.source, forKey: "source")
-                        newIng.setValue(i.isWarning, forKey: "isWarning")
+                        newIng.setValue(ingredient.id, forKey: "id")
+                        newIng.setValue(ingredient.name, forKey: "name")
+                        newIng.setValue(ingredient.composition, forKey: "composition")
+                        newIng.setValue(ingredient.description, forKey: "ing_description")
+                        newIng.setValue(ingredient.source, forKey: "source")
+                        newIng.setValue(ingredient.isWarning, forKey: "isWarning")
                         ingredients.add(newIng)
                     }
                 }
                 let nutritionFacts = newFood.mutableSetValue(forKey: #keyPath(CD_food.nutritionFacts))
-                for n in food.nutritionFacts {
+                for nutritionFact in food.nutritionFacts {
                     if let nfEntity = NSEntityDescription.entity(forEntityName: "CD_nutritionFact", in: context.viewContext) {
                         let newNF = NSManagedObject(entity: nfEntity, insertInto: context.viewContext)
-                        newNF.setValue(n.id, forKey: "id")
-                        newNF.setValue(n.name, forKey: "name")
-                        newNF.setValue(n.unit, forKey: "unit")
-                        newNF.setValue(n.amount, forKey: "amount")
-                        newNF.setValue(n.description, forKey: "nf_description")
-                        newNF.setValue(n.source, forKey: "source")
-                        newNF.setValue(n.isLimiting, forKey: "isLimiting")
+                        newNF.setValue(nutritionFact.id, forKey: "id")
+                        newNF.setValue(nutritionFact.name, forKey: "name")
+                        newNF.setValue(nutritionFact.unit, forKey: "unit")
+                        newNF.setValue(nutritionFact.amount, forKey: "amount")
+                        newNF.setValue(nutritionFact.description, forKey: "nf_description")
+                        newNF.setValue(nutritionFact.source, forKey: "source")
+                        newNF.setValue(nutritionFact.isLimiting, forKey: "isLimiting")
                         nutritionFacts.add(newNF)
                     }
                 }
@@ -92,6 +93,7 @@ class DataManager {
             return false
         }
     }
+    // swiftlint:enable function_body_length
 
     func loadFoods() -> [Food] {
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: "CD_food")
@@ -100,17 +102,17 @@ class DataManager {
         request.returnsObjectsAsFaults = false
         var foods = [Food]()
         do {
-            let result = try context.viewContext.fetch(request)
-            for data in result as! [NSManagedObject] {
+            guard let result = try context.viewContext.fetch(request) as? [NSManagedObject] else { return foods }
+            for data in result {
                 if let id = data.value(forKey: "id") as? UUID,
-                    let api_id = data.value(forKey: "api_id") as? String,
+                    let apiId = data.value(forKey: "api_id") as? String,
                     let upc = data.value(forKey: "upc") as? Int64,
                     let name = data.value(forKey: "name") as? String,
                     let createdAt = data.value(forKey: "created_at") as? Date,
                     let ingredients = decodeIngredients(data.mutableSetValue(forKey: "ingredients") as? NSMutableSet),
                     let nutritionFacts = decodeNutritionFacts(data.mutableSetValue(forKey: "nutritionFacts") as? NSMutableSet)
-                {
-                    let food = Food(id: id, api_id: api_id, upc: upc, name: name, ingredients: ingredients, nutritionFacts: nutritionFacts, createdAt: createdAt)
+                { // swiftlint:disable:this opening_brace
+                    let food = Food(id: id, apiId: apiId, upc: upc, name: name, ingredients: ingredients, nutritionFacts: nutritionFacts, createdAt: createdAt)
                         foods.append(food)
                 }
             }
@@ -119,7 +121,7 @@ class DataManager {
         }
         return foods
     }
-  
+
     private func decodeIngredients(_ dataSet: NSMutableSet?) -> [Ingredient]? {
         var ingredients = [Ingredient]()
         guard let dataSet = dataSet else { return nil }
@@ -128,7 +130,7 @@ class DataManager {
             if let id = data.value(forKey: "id") as? String,
                 let name = data.value(forKey: "name") as? String,
                 let isWarning = data.value(forKey: "isWarning") as? Bool
-            {
+            { // swiftlint:disable:this opening_brace
                 let comp = data.value(forKey: "composition") as? String
                 let desc = data.value(forKey: "ing_description") as? String
                 let source = data.value(forKey: "source") as? String
@@ -149,11 +151,11 @@ class DataManager {
                 let amount = data.value(forKey: "amount") as? Float,
                 let unit = data.value(forKey: "unit") as? String,
                 let isLimiting = data.value(forKey: "isLimiting") as? Bool
-            {
+            { // swiftlint:disable:this opening_brace
                 let desc = data.value(forKey: "nf_description") as? String
                 let source = data.value(forKey: "source") as? String
-                let nf = NutritionFact(id: id, name: name, description: desc, source: source, amount: amount, unit: unit, isLimiting: isLimiting)
-                  nutritionFacts.append(nf)
+                let nutritionFact = NutritionFact(id: id, name: name, description: desc, source: source, amount: amount, unit: unit, isLimiting: isLimiting)
+                  nutritionFacts.append(nutritionFact)
             }
         }
         return nutritionFacts
@@ -164,11 +166,9 @@ class DataManager {
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: "CD_food")
         request.returnsObjectsAsFaults = false
         do {
-          let result = try context.viewContext.fetch(request)
-            for data in result as! [NSManagedObject] {
-                
+            guard let result = try context.viewContext.fetch(request) as? [NSManagedObject] else { return false }
+            for data in result {
                 if let dataId = data.value(forKey: "id") as? UUID {
-            
                     if dataId == id {
                         context.viewContext.delete(data)
                         try context.viewContext.save()
@@ -183,18 +183,28 @@ class DataManager {
     }
 
     func saveSetting(setting: NutrientSetting) -> Bool {
-      if let entity = NSEntityDescription.entity(forEntityName: "CD_nutrientSetting", in: context.viewContext) {
-          let newSetting = NSManagedObject(entity: entity, insertInto: context.viewContext)
-            newSetting.setValue(setting.id, forKey: "id")
-            newSetting.setValue(setting.name, forKey: "name")
-            newSetting.setValue(setting.unit, forKey: "unit")
-            newSetting.setValue(setting.dailyValue, forKey: "dailyValue")
-        }
+        // check if the settings already exists
+        let request = NSFetchRequest<NSFetchRequestResult>(entityName: "CD_nutrientSetting")
+        request.returnsObjectsAsFaults = false
+        request.predicate = NSPredicate(format: "id = %@", setting.id as CVarArg)
         do {
+            let result = try context.viewContext.fetch(request)
+            guard let settingDataList  = result as? [NSManagedObject] else { return false }
+            if settingDataList.count > 0 { // already exists
+                let settingData = settingDataList[0]
+                settingData.setValue(setting.dailyValue, forKey: "dailyValue")
+            } else { // needs to be created
+                if let entity = NSEntityDescription.entity(forEntityName: "CD_nutrientSetting", in: context.viewContext) {
+                    let newSetting = NSManagedObject(entity: entity, insertInto: context.viewContext)
+                    newSetting.setValue(setting.id, forKey: "id")
+                    newSetting.setValue(setting.name, forKey: "name")
+                    newSetting.setValue(setting.unit, forKey: "unit")
+                    newSetting.setValue(setting.dailyValue, forKey: "dailyValue")
+                }
+            }
             try context.viewContext.save()
             return true
         } catch {
-            print("Failed saving setting")
             return false
         }
     }
@@ -204,8 +214,8 @@ class DataManager {
         request.returnsObjectsAsFaults = false
         var settings = [NutrientSetting]()
         do {
-            let result = try context.viewContext.fetch(request)
-            for data in result as! [NSManagedObject] {
+            guard let result = try context.viewContext.fetch(request) as? [NSManagedObject] else { return settings }
+            for data in result {
                 if let id = data.value(forKey: "id") as? UUID,
                     let name = data.value(forKey: "name") as? String,
                     let unit = data.value(forKey: "unit") as? String,
